refactor: /api/auth 폴더(라우터) 생성 - user와 auth 분리
